+++ src/_pyfiberassign.cpp
@@ -1,5 +1,6 @@
 #include <string>
 #include <sstream>
+#include <cstdint>
 
 #include <pybind11/pybind11.h>
 #include <pybind11/operators.h>