+++ doc/changes.rst
@@ -7,12 +7,22 @@ fiberassign change log
 ------------------
 
 * handle dark1b/bright1b (PR `#468`_)
+* Custom too: allow ToO from fiberassign tertiary folder (PR `#476`_)
+* add stuck_on_sky_from_fafns() option (PR `#471`_)
+* get_fba_use_fabs(): add log infos (PR `#478`_)
+* Doc err fix (PR `#479`_)
+* include cstdint for gcc/13 support (PR `#477`_)
 * Freeze downloads of the IERS leap second table in ``fba_launch``, ``fba_run`` and ``fba_rerun`` (PR `#475`_).
 * Field rotation updates to avoid hexapod rotation limit errors (PR `#466`_).
 * Fix C++ bug with use of std::abs and improve floating point reproducibility (PR `#470`_).
 * Remove ``DesiTest`` from setup.py and warn about other deprecated features (PR `#464`_).
 
 .. _`#468`: https://github.com/desihub/fiberassign/pull/468
+.. _`#476`: https://github.com/desihub/fiberassign/pull/476
+.. _`#471`: https://github.com/desihub/fiberassign/pull/471
+.. _`#478`: https://github.com/desihub/fiberassign/pull/478
+.. _`#479`: https://github.com/desihub/fiberassign/pull/479
+.. _`#477`: https://github.com/desihub/fiberassign/pull/477
 .. _`#475`: https://github.com/desihub/fiberassign/pull/475
 .. _`#466`: https://github.com/desihub/fiberassign/pull/466
 .. _`#470`: https://github.com/desihub/fiberassign/pull/470